実行時間制限を問題文から取得、ローカルテストに反映する機能を追加 (#254)
* timeoutを問題文から取得するのを追加,
testerのMetadata完全一致を修正

* テストのmetadataをtimeoutに対応

* metadataから制限時間を読むテストを追加

* timeout->timeout_ms

* READMEのtimeout_adjustmentの記述を修正。timeout_adjustmentのテストを追加

+++ atcodertools/executils/run_program.py
@@ -30,7 +30,7 @@ class ExecResult:
         return len(self.stderr) > 0
 
 
-def run_program(exec_file: str, input_file: str, timeout_sec: int, args=None, current_working_dir: str = None) -> ExecResult:
+def run_program(exec_file: str, input_file: str, timeout_sec: float, args=None, current_working_dir: str = None) -> ExecResult:
     if args is None:
         args = []
     try:
